003-cost-alerts/requirements.txt
goblet-gcp==0.4.9
slack_sdk==3.5.0
google-cloud-bigquery==2.16.0